project/api/requirements.txt
fastapi
uvicorn
onnxruntime
numpy
pillow
torchvision
